backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2
pgvector==0.3.5

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.2
openpyxl==3.1.5

# AI / Embedding
httpx==0.27.2

# Utils
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
uuid6==2024.7.10
